leo21

+++ src/test/java/leo14/parser/intParserTest.kt
@@ -2,6 +2,7 @@ package leo14.parser
 
 import leo.base.assertEqualTo
 import leo14.number
+import java.math.BigDecimal
 import kotlin.test.Test
 import kotlin.test.assertFails
 
@@ -16,9 +17,9 @@ class NumberParserTest {
 		parseNumber("-01").assertEqualTo(number(-1))
 		parseNumber("-012").assertEqualTo(number(-12))
 
-		parseNumber("0.0").assertEqualTo(number(0.0))
+		parseNumber("0.0").assertEqualTo(number(BigDecimal.valueOf(0.0)))
 		parseNumber("0.1").assertEqualTo(number(0.1))
-		parseNumber("1.0").assertEqualTo(number(1.0))
+		parseNumber("1.0").assertEqualTo(number(BigDecimal.valueOf(1.0)))
 		parseNumber("1.1").assertEqualTo(number(1.1))
 		parseNumber("3.1415").assertEqualTo(number(3.1415))
 		parseNumber("123.456").assertEqualTo(number(123.456))
